requirements.txt
asgiref==3.8.1
Django==5.1.7
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
phonenumbers==9.0.1
sqlparse==0.5.3
tzdata==2025.2
